Fixed static analysis issues
